gcp/requirements.txt
tensorflow==2.12.0
google-cloud-storage==2.10.0
Pillow==10.0.0
functions-framework==3.5.0